+++ .gitignore
@@ -23,6 +23,7 @@ dist-ssr
 *.sln
 *.sw?
 
-# .gitignore
+# sto mi netreba
 fire
 .env
+unused
