+++ package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.1",
     "react-dom": "^18.2.0",
+    "react-firebase-hooks": "^5.1.1",
     "react-router-dom": "^6.21.1",
     "react-toastify": "^9.1.3"
   },
