--- src/Cart.jsx
@@ -1,41 +1,0 @@
-import React, { useState } from 'react';
-import { Button, Modal } from 'react-bootstrap';
-
-//modal ovo radi pa cemo to ostaviti zasada tako
-// backup verzija
-
-const ShoppingCart = () => {
-  const [showModal, setShowModal] = useState(false);
-
-  const handleShow = () => setShowModal(true);
-  const handleClose = () => setShowModal(false);
-
-  // ono iz pizzaItem tj shopping cart modal 
-  //samo useShopping cart mogu ubaciti ovdje
-  return (
-    <div>
-      <Button variant="dark" className="rounded bg-danger" onClick={handleShow}>
-        Check Cart <i class="bi bi-cart4"></i>
-      </Button>
-
-      <Modal show={showModal} onHide={handleClose}>
-        <Modal.Header closeButton>
-          <Modal.Title>Shopping Cart</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          {/* Add your shopping cart content here */}
-          <p>Item 1: Product Name</p>
-          <p>Item 2: Another Product</p>
-          {/* Add more items as needed */}
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Close
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </div>
-  );
-};
-
-export default ShoppingCart;
