--- src/mystyle.module.css
@@ -1,38 +1,0 @@
-.darkTheme {
-    background-color: #2a2727;
-    color: white;
-    font-family: Arial, sans-serif;
-    font-weight: bold;
-}
-
-ul {
-    color: white;
-    text-align: left;
-    font-weight: lighter;
-}
-
-li {
-    font-size: 16px;
-    color:white;
-}
-
-button {
-    float: right;
-}
-
-h5 {
-    text-align: left;
-    font-size: 22px;
-    font-weight: bold;
-    padding-top: 10px;
-    padding-left: 20px;
-    font-family: Georgia, serif;
-}
-
-nav {
-    font-family: Georgia, serif;
-    font-size: 20px;
-}
-
-/* nisam targeto nav elemente sa klasom nav-item jer sam ih vec targeto sa ul klasom 
-jer koristim to za sve i za menu i za nav*/
